Keep Pierre off the app boot path
useAppTheme is eager. Registering custom code themes from that hook pulled
@pierre/diffs into the boot payload and blew the bundle budget. Move
registration next to the worker-pool sync, which already loads Pierre.
