+++ apps/app/src/lib/code-theme.ts
@@ -1,9 +1,6 @@
-import { useEffect, useMemo, useSyncExternalStore } from "react";
-import { registerCustomTheme } from "@pierre/diffs";
-import { useWorkerPool } from "@pierre/diffs/react";
+import { useMemo, useSyncExternalStore } from "react";
 import {
   defaultResolvedCodeTheme,
-  stampRegisteredThemeName,
   type ResolvedCodeTheme,
 } from "@bb/domain";
 
@@ -12,7 +9,6 @@ const CODE_THEME_LIGHT_DATASET = "bbCodeThemeLight";
 
 let currentResolvedCodeTheme: ResolvedCodeTheme = defaultResolvedCodeTheme;
 const subscribers = new Set<() => void>();
-const registeredFileNames = new Set<string>();
 
 function publish(): void {
   for (const subscriber of subscribers) subscriber();
@@ -23,15 +19,6 @@ function writeDocumentDataset(resolved: ResolvedCodeTheme): void {
   const root = document.documentElement;
   root.dataset[CODE_THEME_DARK_DATASET] = resolved.dark;
   root.dataset[CODE_THEME_LIGHT_DATASET] = resolved.light;
-}
-
-function registerFiles(resolved: ResolvedCodeTheme): void {
-  for (const [name, theme] of Object.entries(resolved.files)) {
-    if (registeredFileNames.has(name)) continue;
-    registeredFileNames.add(name);
-    const stamped = stampRegisteredThemeName(name, theme);
-    registerCustomTheme(name, () => Promise.resolve(stamped));
-  }
 }
 
 export function getResolvedCodeTheme(): ResolvedCodeTheme {
@@ -46,11 +33,11 @@ export function subscribeResolvedCodeTheme(callback: () => void): () => void {
 }
 
 /**
- * Register custom Pierre JSON files and publish the resolved dark/light names
- * for every FileDiff / File surface and first-party plugin renderers.
+ * Publish the resolved dark/light names for every FileDiff / File surface
+ * and first-party plugin renderers. Pierre file registration lives next to
+ * the worker-pool sync so `@pierre/diffs` stays off the app boot path.
  */
 export function applyResolvedCodeTheme(resolved: ResolvedCodeTheme): void {
-  registerFiles(resolved);
   writeDocumentDataset(resolved);
   if (
     currentResolvedCodeTheme.dark === resolved.dark &&
@@ -80,22 +67,6 @@ export function useResolvedCodeThemePair(): {
     () => ({ dark: resolved.dark, light: resolved.light }),
     [resolved.dark, resolved.light],
   );
-}
-
-/**
- * File / FileDiff ignore `options.theme` while a worker pool is active and
- * highlight with the pool's render options instead. Push the resolved pair
- * into that pool after custom JSON is registered.
- */
-export function useSyncPierreWorkerPoolTheme(): void {
-  const pool = useWorkerPool();
-  const theme = useResolvedCodeThemePair();
-  useEffect(() => {
-    if (pool == null) return;
-    void pool.setRenderOptions({ theme }).catch((error: unknown) => {
-      console.error("Failed to apply the code theme to the Pierre worker pool", error);
-    });
-  }, [pool, theme]);
 }
 
 /** Read the host-published names. Plugins that render FileDiff should use this. */
